[harvester] Handle multiple resource locators
We now create as many package resources as resource locators are present
in the XML document. The following information is extracted (if present):
name
description
url
function
protocol

format is set to 'WMS' if it's a verified WMS, otherwise 'Unverified'.
If WMS are verified, a 'verified' and 'verified_date' are added. Also
a 'ckan_recommended_wms_preview' is added.